Move TopLevelNavRoot to top level of compose package
